track the search state
